rename Stash to BitbucketServer
