+++ master/buildbot/test/unit/test_reporter_bitbucketserver.py
@@ -25,15 +25,15 @@ from buildbot import config
 from buildbot.process.properties import Interpolate
 from buildbot.process.results import FAILURE
 from buildbot.process.results import SUCCESS
-from buildbot.reporters.stash import StashStatusPush
-from buildbot.reporters.stash import StashPRCommentPush
+from buildbot.reporters.bitbucketserver import BitbucketServerPRCommentPush
+from buildbot.reporters.bitbucketserver import BitbucketServerStatusPush
 from buildbot.test.fake import httpclientservice as fakehttpclientservice
 from buildbot.test.fake import fakemaster
 from buildbot.test.util.logging import LoggingMixin
 from buildbot.test.util.reporter import ReporterTestMixin
 
 
-class TestStashStatusPush(unittest.TestCase, ReporterTestMixin, LoggingMixin):
+class TestBitbucketServerStatusPush(unittest.TestCase, ReporterTestMixin, LoggingMixin):
 
     @defer.inlineCallbacks
     def setupReporter(self, **kwargs):
@@ -46,7 +46,7 @@ class TestStashStatusPush(unittest.TestCase, ReporterTestMixin, LoggingMixin):
             self.master, self,
             'serv', auth=('username', 'passwd'),
             debug=None, verify=None)
-        self.sp = sp = StashStatusPush("serv", "username", "passwd", **kwargs)
+        self.sp = sp = BitbucketServerStatusPush("serv", "username", "passwd", **kwargs)
         yield sp.setServiceParent(self.master)
         yield self.master.startService()
 
@@ -139,10 +139,10 @@ class TestStashStatusPush(unittest.TestCase, ReporterTestMixin, LoggingMixin):
         build['complete'] = False
         self.setUpLogging()
         self.sp.buildStarted(("build", 20, "started"), build)
-        self.assertLogged('404: Unable to send Stash status')
+        self.assertLogged('404: Unable to send Bitbucket Server status')
 
 
-class TestStashPRCommentPush(unittest.TestCase, ReporterTestMixin, LoggingMixin):
+class TestBitbucketServerPRCommentPush(unittest.TestCase, ReporterTestMixin, LoggingMixin):
 
     @defer.inlineCallbacks
     def setupReporter(self, **kwargs):
@@ -154,8 +154,8 @@ class TestStashPRCommentPush(unittest.TestCase, ReporterTestMixin, LoggingMixin)
         self._http = yield fakehttpclientservice.HTTPClientService.getFakeService(
             self.master, self, 'serv', auth=('username', 'passwd'), debug=None,
             verify=None)
-        self.sp = sp = StashPRCommentPush("serv", "username", "passwd", **kwargs)
-        yield sp.setServiceParent(self.master)
+        self.cp = cp = BitbucketServerPRCommentPush("serv", "username", "passwd", **kwargs)
+        yield cp.setServiceParent(self.master)
         yield self.master.startService()
 
     @defer.inlineCallbacks
@@ -182,17 +182,17 @@ class TestStashPRCommentPush(unittest.TestCase, ReporterTestMixin, LoggingMixin)
             code=201)
         build["complete"] = False
         # this shouldn't send anything
-        self.sp.buildStarted(("build", 20, "started"), build)
+        self.cp.buildStarted(("build", 20, "started"), build)
         self.assertEqual(len(self._http._expected), 1)
         build["complete"] = True
-        self.sp.buildFinished(("build", 20, "finished"), build)
+        self.cp.buildFinished(("build", 20, "finished"), build)
         self._http.expect(
             "post",
             u'/rest/api/1.0/projects/PRO/repos/myrepo/pull-requests/20/comments',
             json={"text": "Builder: Builder0 Status: FAILED"},
             code=201)
         build["results"] = FAILURE
-        self.sp.buildFinished(("build", 20, "finished"), build)
+        self.cp.buildFinished(("build", 20, "finished"), build)
 
     @defer.inlineCallbacks
     def test_setting_options(self):
@@ -204,17 +204,17 @@ class TestStashPRCommentPush(unittest.TestCase, ReporterTestMixin, LoggingMixin)
             json={'text': 'url: http://localhost:8080/#builders/79/builds/0 status: SUCCESS'},
             code=201)
         build["complete"] = False
-        self.sp.buildStarted(("build", 20, "started"), build)
+        self.cp.buildStarted(("build", 20, "started"), build)
         self.assertEqual(len(self._http._expected), 1)
         build["complete"] = True
-        self.sp.buildFinished(("build", 20, "finished"), build)
+        self.cp.buildFinished(("build", 20, "finished"), build)
         self._http.expect(
             "post",
             u'/rest/api/1.0/projects/PRO/repos/myrepo/pull-requests/20/comments',
             json={'text': 'url: http://localhost:8080/#builders/79/builds/0 status: FAILED'},
             code=201)
         build["results"] = FAILURE
-        self.sp.buildFinished(("build", 20, "finished"), build)
+        self.cp.buildFinished(("build", 20, "finished"), build)
 
     @defer.inlineCallbacks
     def test_error(self):
@@ -228,5 +228,5 @@ class TestStashPRCommentPush(unittest.TestCase, ReporterTestMixin, LoggingMixin)
             content_json=None)
         self.setUpLogging()
         build['complete'] = True
-        self.sp.buildFinished(("build", 20, "finished"), build)
+        self.cp.buildFinished(("build", 20, "finished"), build)
         self.assertLogged('404: Unable to send a comment: None')
